refactor: move user statistics service
