+++ app/commands/users/calculate_statistics.rb
@@ -34,7 +34,7 @@ module Users
     private
 
     def user_statistics_service
-      @user_statistics_service ||= Service::Users::Statistics.new(donations:, user:, customer:)
+      @user_statistics_service ||= UserServices::Statistics.new(donations:, user:, customer:)
     end
 
     def crypto_user_statistics_service
